ref: accelerator connector methods x/n (#3469)
* ref: accelerator connector methods x/n

* ref: accelerator connector methods x/n
