+++ pytorch_lightning/accelerators/accelerator_connector.py
@@ -2,7 +2,25 @@ from pytorch_lightning import accelerators
 import os
 import torch
 from pytorch_lightning.utilities import device_parser
-from pytorch_lightning.utilities import rank_zero_warn, rank_zero_only
+from pytorch_lightning.utilities import rank_zero_only
+from pytorch_lightning.utilities.distributed import rank_zero_warn, rank_zero_info
+from pytorch_lightning.utilities.exceptions import MisconfigurationException
+from pytorch_lightning import _logger as log
+import re
+
+try:
+    import torch_xla
+except ImportError:
+    XLA_AVAILABLE = False
+else:
+    XLA_AVAILABLE = True
+
+try:
+    import horovod.torch as hvd
+except (ModuleNotFoundError, ImportError):
+    HOROVOD_AVAILABLE = False
+else:
+    HOROVOD_AVAILABLE = True
 
 
 class AcceleratorConnector:
@@ -79,23 +97,23 @@ class AcceleratorConnector:
 
         # distributed backend choice
         self.trainer.distributed_backend = distributed_backend
-        self.trainer.set_distributed_mode(distributed_backend)
+        self.set_distributed_mode(distributed_backend)
 
         # override dist backend when using tpus
         if self.trainer.on_tpu:
             self.trainer.distributed_backend = 'tpu'
-            self.trainer.init_tpu()
+            self.trainer.use_tpu = True
 
         # init flags for SLURM+DDP to work
         self.trainer.world_size = 1
         self.trainer.interactive_ddp_procs = []
-        self.trainer.configure_slurm_ddp(self.trainer.num_nodes)
-        self.trainer.node_rank = self.trainer.determine_ddp_node_rank()
-        self.trainer.local_rank = self.trainer.determine_local_rank()
+        self.configure_slurm_ddp(self.trainer.num_nodes)
+        self.trainer.node_rank = self.determine_ddp_node_rank()
+        self.trainer.local_rank = self.determine_local_rank()
         self.trainer.global_rank = 0
 
         # NVIDIA setup
-        self.trainer.set_nvidia_flags(self.trainer.is_slurm_managing_tasks, self.trainer.data_parallel_device_ids)
+        self.set_nvidia_flags(self.trainer.is_slurm_managing_tasks, self.trainer.data_parallel_device_ids)
 
         self.trainer.on_colab_kaggle = os.getenv('COLAB_GPU') or os.getenv('KAGGLE_URL_BASE')
 
@@ -143,3 +161,192 @@ class AcceleratorConnector:
             accelerator_backend = accelerators.CPUBackend(self.trainer)
 
         return accelerator_backend
+
+    def set_distributed_mode(self, distributed_backend):
+        self.trainer.use_dp = False
+        self.trainer.use_ddp = False
+        self.trainer.use_ddp2 = False
+        self.trainer.use_horovod = False
+        self.trainer.use_single_gpu = False
+
+        if distributed_backend is None:
+            if self.has_horovodrun():
+                self._set_horovod_backend()
+            elif self.trainer.num_gpus == 0:
+                if self.trainer.num_nodes > 1 or self.trainer.num_processes > 1:
+                    self.trainer.use_ddp = True  # ddp_cpu
+            elif self.trainer.num_gpus == 1:
+                self.trainer.use_single_gpu = True
+            elif self.trainer.num_gpus > 1:
+                rank_zero_warn(
+                    'You requested multiple GPUs but did not specify a backend, e.g.'
+                    ' Trainer(distributed_backend=dp) (or ddp, ddp2).'
+                    ' Setting distributed_backend=ddp_spawn for you.'
+                )
+                self.trainer.distributed_backend = 'ddp_spawn'
+                distributed_backend = 'ddp_spawn'
+
+        if distributed_backend == "dp":
+            # do nothing if num_gpus == 0
+            if self.trainer.num_gpus == 1:
+                self.trainer.use_single_gpu = True
+                self.trainer.use_dp = True
+            elif self.trainer.num_gpus > 1:
+                self.trainer.use_dp = True
+
+        elif distributed_backend in ['ddp', 'ddp_spawn']:
+            if self.trainer.num_gpus == 0:
+                if self.trainer.num_nodes > 1 or self.trainer.num_processes > 1:
+                    self.trainer.use_ddp = True  # ddp_cpu
+            elif self.trainer.num_gpus == 1:
+                self.trainer.use_single_gpu = True
+                self.trainer.use_ddp = True
+            elif self.trainer.num_gpus > 1:
+                self.trainer.use_ddp = True
+                self.trainer.num_processes = self.trainer.num_gpus
+
+        elif distributed_backend == "ddp2":
+            # do nothing if num_gpus == 0
+            if self.trainer.num_gpus >= 1:
+                self.trainer.use_ddp2 = True
+        elif distributed_backend == "ddp_cpu":
+            if self.trainer.num_gpus > 0:
+                rank_zero_warn(
+                    'You requested one or more GPUs, but set the backend to `ddp_cpu`. Training will not use GPUs.'
+                )
+            self.trainer.use_ddp = True
+            self.trainer.data_parallel_device_ids = None
+            self.trainer.on_gpu = False
+        elif distributed_backend == 'horovod':
+            self._set_horovod_backend()
+
+        # throw error to force user ddp or ddp2 choice
+        if self.trainer.num_nodes > 1 and not (self.trainer.use_ddp2 or self.trainer.use_ddp):
+            raise MisconfigurationException(
+                'DataParallel does not support num_nodes > 1. Switching to DistributedDataParallel for you. '
+                'To silence this warning set distributed_backend=ddp or distributed_backend=ddp2'
+            )
+
+        rank_zero_info(f'GPU available: {torch.cuda.is_available()}, used: {self.trainer.on_gpu}')
+        num_cores = self.trainer.tpu_cores if self.trainer.tpu_cores is not None else 0
+        rank_zero_info(f'TPU available: {XLA_AVAILABLE}, using: {num_cores} TPU cores')
+
+        if torch.cuda.is_available() and not self.trainer.on_gpu:
+            rank_zero_warn('GPU available but not used. Set the --gpus flag when calling the script.')
+
+    def _set_horovod_backend(self):
+        self.check_horovod()
+        self.trainer.use_horovod = True
+
+        # Initialize Horovod to get rank / size info
+        hvd.init()
+        if self.trainer.on_gpu:
+            # Horovod assigns one local GPU per process
+            self.trainer.root_gpu = hvd.local_rank()
+
+    def check_horovod(self):
+        """Raises a `MisconfigurationException` if the Trainer is not configured correctly for Horovod."""
+        if not HOROVOD_AVAILABLE:
+            raise MisconfigurationException(
+                'Requested `distributed_backend="horovod"`, but Horovod is not installed.'
+                'Install with \n $HOROVOD_WITH_PYTORCH=1 pip install horovod[pytorch]'
+            )
+
+        if self.trainer.num_gpus > 1 or self.trainer.num_nodes > 1:
+            raise MisconfigurationException(
+                'Horovod does not support setting num_nodes / num_gpus explicitly. Use '
+                'horovodrun / mpirun to configure the number of processes.'
+            )
+
+    @staticmethod
+    def has_horovodrun():
+        """Returns True if running with `horovodrun` using Gloo or OpenMPI."""
+        return 'OMPI_COMM_WORLD_RANK' in os.environ or 'HOROVOD_RANK' in os.environ
+
+    def set_nvidia_flags(self, is_slurm_managing_tasks, data_parallel_device_ids):
+        if data_parallel_device_ids is None:
+            return
+
+        # set the correct cuda visible devices (using pci order)
+        os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
+
+        # when slurm is managing the task it sets the visible devices
+        if not is_slurm_managing_tasks and 'CUDA_VISIBLE_DEVICES' not in os.environ:
+            if isinstance(data_parallel_device_ids, int):
+                id_str = ','.join(str(x) for x in list(range(data_parallel_device_ids)))
+                os.environ["CUDA_VISIBLE_DEVICES"] = id_str
+            else:
+                gpu_str = ','.join([str(x) for x in data_parallel_device_ids])
+                os.environ["CUDA_VISIBLE_DEVICES"] = gpu_str
+
+        # don't make this debug... this is good UX
+        rank_zero_info(f'CUDA_VISIBLE_DEVICES: [{os.environ["CUDA_VISIBLE_DEVICES"]}]')
+
+    def determine_local_rank(self):
+        if self.trainer.is_slurm_managing_tasks:
+            return int(os.environ['SLURM_LOCALID'])
+        else:
+            return int(os.environ.get('LOCAL_RANK', 0))
+
+    def configure_slurm_ddp(self, num_gpu_nodes):
+        self.trainer.is_slurm_managing_tasks = False
+
+        # extract SLURM flag vars
+        # whenever we have the correct number of tasks, we let slurm manage processes
+        # otherwise we launch the required number of processes
+        if self.trainer.use_ddp:
+            self.trainer.num_requested_gpus = self.trainer.num_gpus * num_gpu_nodes
+            self.trainer.num_slurm_tasks = 0
+            try:
+                self.trainer.num_slurm_tasks = int(os.environ['SLURM_NTASKS'])
+                self.trainer.is_slurm_managing_tasks = self.trainer.num_slurm_tasks == self.trainer.num_requested_gpus
+
+                # in interactive mode we don't manage tasks
+                job_name = os.environ['SLURM_JOB_NAME']
+                if job_name == 'bash':
+                    self.trainer.is_slurm_managing_tasks = False
+
+            except Exception:
+                # likely not on slurm, so set the slurm managed flag to false
+                self.trainer.is_slurm_managing_tasks = False
+
+        # used for tests only, set this flag to simulate slurm managing a task
+        try:
+            should_fake = int(os.environ['FAKE_SLURM_MANAGING_TASKS'])
+            if should_fake:
+                self.trainer.is_slurm_managing_tasks = True
+        except Exception:
+            pass
+
+        # notify user the that slurm is managing tasks
+        if self.trainer.is_slurm_managing_tasks:
+            rank_zero_info('Multi-processing is handled by Slurm.')
+
+    def determine_ddp_node_rank(self):
+        if self.trainer.is_slurm_managing_tasks:
+            return int(os.environ['SLURM_NODEID'])
+
+        # torchelastic uses the envvar GROUP_RANK, whereas other systems(?) use NODE_RANK.
+        # otherwise use given node rank or default to node rank 0
+        env_vars = ['NODE_RANK', 'GROUP_RANK']
+        node_ids = [(k, os.environ.get(k, None)) for k in env_vars]
+        node_ids = [(k, v) for k, v in node_ids if v is not None]
+        if len(node_ids) == 0:
+            return 0
+        if len(node_ids) > 1:
+            log.warning(f"Multiple environment variables ({node_ids}) defined for node rank. Using the first one.")
+        k, rank = node_ids.pop()
+        rank_zero_info(f"Using environment variable {k} for node rank ({rank}).")
+        return int(rank)
+
+    def resolve_root_node_address(self, root_node):
+        if '[' in root_node:
+            name, numbers = root_node.split('[', maxsplit=1)
+            number = numbers.split(',', maxsplit=1)[0]
+            if '-' in number:
+                number = number.split('-')[0]
+
+            number = re.sub('[^0-9]', '', number)
+            root_node = name + number
+
+        return root_node
